requirements.txt
# Core dependencies
aiohttp==3.9.1
aiofiles==23.2.1
asyncio==3.4.3

# Web interface
aiohttp-jinja2==1.5.1
jinja2==3.1.2

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.7
certifi==2023.11.17

# Configuration and serialization
PyYAML==6.0.1
pydantic==2.5.2

# Monitoring and metrics
psutil==5.9.6

# Data validation
jsonschema==4.20.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation (optional)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0